auth ready for production

+++ controllers/studiesController.js
@@ -1,18 +1,14 @@
 const queries = require('../DB/StudyQueries/studyQueries');
 const {APIError} = require('../ErrorHaneling/APIError')
-const {unAuthorized} =require('../Auth/authUtilis')
-const getAllStudies = async (req, res, next) => {
-    try {
-        res.status(200).json(await queries.getAllStudies());
-    } catch (err) {
-        const apiError = new APIError(err)
-        next(apiError, req, res);
-    }
-}
+const {unAuthorized} = require('../Auth/authUtilis')
+
+
 const getStudyById = async (req, res, next) => {
     try {
+        const user = req.user;
+        if (!user) return await unAuthorized(req, res)
         const studyId = req.params.id;
-        res.status(200).json(await queries.getStudyById(studyId));
+        res.status(200).json(await queries.getStudyById(user.researcher_id, studyId));
     } catch (err) {
         const apiError = new APIError(err)
         next(apiError, req, res);
@@ -22,8 +18,12 @@ const getStudyById = async (req, res, next) => {
 
 const createStudy = async (req, res, next) => {
     try {
+        const user = req.user;
+        if (!user) return await unAuthorized(req, res)
         const {study_name, study_subject, study_prompt, study_description} = req.body;
-        res.status(200).json(await queries.createStudy(study_name,study_subject,study_prompt,study_description));
+        const row = await queries.createStudy(study_name, study_subject, study_prompt, study_description);
+        await queries.joinStudyToResearcher(user.researcher_id, row.study_id);
+        res.status(200).json(row);
     } catch (err) {
         const apiError = new APIError(err)
         next(apiError, req, res);
@@ -33,6 +33,8 @@ const createStudy = async (req, res, next) => {
 
 const updateStudy = async (req, res, next) => {
     try {
+        const user = req.user;
+        if (!user) return await unAuthorized(req, res)
         const {study_id} = req.body;
         if (!study_id) throw new Error('study_id required in request body for update');
         delete req.body.study_id;
@@ -45,9 +47,10 @@ const updateStudy = async (req, res, next) => {
 }
 
 
-
 const deleteStudy = async (req, res, next) => {
     try {
+        const user = req.user;
+        if (!user) return await unAuthorized(req, res)
         const {study_id} = req.body;
         await queries.deleteStudy(study_id);
         res.status(200)
@@ -59,6 +62,8 @@ const deleteStudy = async (req, res, next) => {
 
 const getExperimentsByStudyId = async (req, res, next) => {
     try {
+        const user = req.user;
+        if (!user) return await unAuthorized(req, res)
         const studyId = req.params.id;
         res.status(200).json(await queries.getExperimentsByStudyId(studyId));
     } catch (err) {
@@ -67,11 +72,10 @@ const getExperimentsByStudyId = async (req, res, next) => {
     }
 }
 
-const getStudiesByResearcherId = async (req,res,next) => {
+const getStudiesByResearcherId = async (req, res, next) => {
     try {
         const user = req.user;
-        if (!user) return await unAuthorized(req,res)
-        console.log(user)
+        if (!user) return await unAuthorized(req, res)
         res.status(200).json(await queries.getStudiesByResearcherId(user.researcher_id));
     } catch (err) {
         const apiError = new APIError(err)
@@ -81,7 +85,6 @@ const getStudiesByResearcherId = async (req,res,next) => {
 
 
 module.exports = {
-    getAllStudies,
     getStudyById,
     createStudy,
     updateStudy,
